+++ axios.js
@@ -3,8 +3,6 @@ import { wrapper as axiosCookieJarSupport } from 'axios-cookiejar-support';
 import tough from 'tough-cookie';
 import { Session } from './core.js';
 
-const defaultUserAgent = 'm3api/0.3.0 (https://www.npmjs.com/package/m3api)';
-
 function transformResponse( response ) {
 	const headers = { ...response.headers };
 	delete headers[ 'set-cookie' ];
@@ -18,17 +16,10 @@ function transformResponse( response ) {
 
 class AxiosSession extends Session {
 
-	constructor( apiUrl, defaultParams = {}, userAgent = '' ) {
-		super( apiUrl, defaultParams, userAgent );
+	constructor( apiUrl, defaultParams = {}, defaultOptions = {} ) {
+		super( apiUrl, defaultParams, defaultOptions );
 		this.session = axios.create( {
 			baseURL: apiUrl,
-			headers: {
-				common: {
-					'user-agent': userAgent ?
-						`${userAgent} ${defaultUserAgent}` :
-						defaultUserAgent,
-				},
-			},
 		} );
 		axiosCookieJarSupport( this.session );
 		Object.assign( this.session.defaults, { // must happen after axiosCookieJarSupport
@@ -36,19 +27,25 @@ class AxiosSession extends Session {
 		} );
 	}
 
-	async internalGet( params ) {
+	async internalGet( params, userAgent ) {
 		const response = await this.session.request( {
 			method: 'GET',
 			params,
+			headers: {
+				'user-agent': userAgent,
+			},
 		} );
 		return transformResponse( response );
 	}
 
-	async internalPost( urlParams, bodyParams ) {
+	async internalPost( urlParams, bodyParams, userAgent ) {
 		const response = await this.session.request( {
 			method: 'POST',
 			params: urlParams,
 			data: new URLSearchParams( bodyParams ),
+			headers: {
+				'user-agent': userAgent,
+			},
 		} );
 		return transformResponse( response );
 	}