+++ fetch.js
@@ -2,8 +2,6 @@
 
 import { Session } from './core.js';
 
-const defaultUserAgent = 'm3api/0.3.0 (https://www.npmjs.com/package/m3api)';
-
 async function transformResponse( response ) {
 	const headers = {};
 	for ( const [ name, value ] of response.headers.entries() ) {
@@ -19,32 +17,31 @@ async function transformResponse( response ) {
 
 class FetchSession extends Session {
 
-	constructor( apiUrl, defaultParams = {}, userAgent = '' ) {
-		super( apiUrl, defaultParams, userAgent );
+	constructor( apiUrl, defaultParams = {}, defaultOptions = {} ) {
+		super( apiUrl, defaultParams, defaultOptions );
 		this.baseUrl = apiUrl;
-		this.headers = {
-			'api-user-agent': userAgent ?
-				`${userAgent} ${defaultUserAgent}` :
-				defaultUserAgent,
-		};
 	}
 
-	async internalGet( params ) {
+	async internalGet( params, userAgent ) {
 		const url = new URL( this.baseUrl );
 		url.search = new URLSearchParams( params );
 		const response = await fetch( url, {
-			headers: this.headers,
+			headers: {
+				'api-user-agent': userAgent,
+			},
 		} );
 		return transformResponse( response );
 	}
 
-	async internalPost( urlParams, bodyParams ) {
+	async internalPost( urlParams, bodyParams, userAgent ) {
 		const url = new URL( this.baseUrl );
 		url.search = new URLSearchParams( urlParams );
 		const response = await fetch( url, {
 			method: 'POST',
 			body: new URLSearchParams( bodyParams ),
-			headers: this.headers,
+			headers: {
+				'api-user-agent': userAgent,
+			},
 		} );
 		return transformResponse( response );
 	}